storage/pyproject.toml
[tool.poetry]
name = "storage"
version = "0.1.0"
description = ""
authors = [
    "IgnisDa <ignisda2002@gmail.com>",
]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
validate_email = "^1.3"
Django = {extras = ["argon2"], version = "^3.1.4"}
psycopg2-binary = "^2.8.6"
django-cors-headers = "^3.6.0"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
pytest-cov = "^2.10.1"
pytest-flake8 = "^1.0.6"
pytest = "^6.1.2"
flake8 = "^3.8.4"
commitizen = "^2.10.0"
pytest-django = "^4.1.0"
Faker = "^5.0.1"
requests = "^2.25.0"
django-extensions = "^3.1.0"
httpie = "^2.3.0"
black = {version = "^20.8b1", allow-prereleases = true}
mypy = "^0.800"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 89
